rename circuit for tester
